feat(courrier-service): reorganize folder structure
