+++ src/use-case/delivery-cost/delivery-cost.service.spec.ts
@@ -2,13 +2,13 @@ import 'reflect-metadata'
 import { jest } from '@jest/globals'
 import { mock, mockClear } from 'jest-mock-extended'
 import { DeliveryCostService } from './delivery-cost.service'
-import { container } from '../container'
-import { TYPES } from '../types'
-import { OfferRepository } from '../domain/repository'
-import { InquireService } from './inquire.service'
-import { CliTable } from '../libs/cli-table'
-import { DiscountType } from '../domain/offer'
 import Table from 'cli-table3'
+import { container } from '../../container'
+import { DiscountType } from '../../domain/offer'
+import { OfferRepository } from '../../domain/repository'
+import { CliTable } from '../../libs/cli-table'
+import { TYPES } from '../../types'
+import { InquireService } from '../inquire/inquire.service'
 
 const mockedDeliveryCostService = mock<DeliveryCostService>()
 const mockedOfferRepository = mock<OfferRepository>()
